Added Heap class, ... pt 2

+++ smash.cpp
@@ -60,32 +60,63 @@ namespace SignalHandlers {
     }
 
     void alarmHandler(int sig_num) {
+        // do we need to print we got an alarm anyway?
         cout << "smash: got an alarm" << endl;
-
-        //find command that cause alarm
-        ProcessControlBlock* lateProcess = shell->getLateProcess();
-
-        //send SIGKILL
-        if (kill(lateProcess->getProcessId(), SIGKILL) < 0) {
-            cerr << "smash error: kill failed" << endl;
+        ProcessControlBlock* lateProcess;
+        //in case of foreground process
+        if ((shell->getIsForgroundTimed()) &&
+        (difftime(time(nullptr), shell->getForegroundProcess()->getStartTime()) == shell->getForegroundProcess()->duration)){
+        lateProcess = const_cast<ProcessControlBlock*>(shell->getForegroundProcess());
+        shell->setIsForgroundTimed(false);
+            //send SIGKILL
+            if (kill(lateProcess->getProcessId(), SIGKILL) < 0) {
+                DEBUG_PRINT("alarmhandler kill failed 1");
+                cerr << "smash error: kill failed" << endl;
+                return;
+            }
+            cout << "smash: " << lateProcess->getCreatingCommand() << " timed out!" << endl;
             return;
         }
 
-        cout << "smash: " << lateProcess->getCreatingCommand() << " timed out!" << endl;
+
+        //find command that cause alarm
+        else lateProcess = shell->getLateProcess();
+
+        //send SIGKILL
+        if (lateProcess) {
+            if (kill(lateProcess->getProcessId(), SIGKILL) < 0) {
+                cerr << "smash error: kill failed" << endl;
+                return;
+            }
+
+            cout << "smash: " << lateProcess->getCreatingCommand() << " timed out!" << endl;
+            // general remove from job list
+            //shell->jobs.removeJobById(lateProcess->getJobId());
+            shell->RemoveLateProcess(lateProcess->getJobId());
+        }
     }
 }
 
-int main(int argc, char* argv[]) {
-    DEBUG_PRINT("this pid is " << getpid()<<endl);
-    if(signal(SIGTSTP , SignalHandlers::ctrlZHandler)==SIG_ERR) {
-        perror("smash error: failed to set ctrl-Z handler");
-    }
-    if(signal(SIGINT , SignalHandlers::ctrlCHandler)==SIG_ERR) {
-        perror("smash error: failed to set ctrl-C handler");
-    }
-    if(signal(SIGALRM , SignalHandlers::alarmHandler)==SIG_ERR) {
-        perror("smash error: failed to set alarm signal handler");
-    }
+    int main(int argc, char *argv[]) {
+        //set sigaction struct
+
+        struct sigaction action;
+        action.sa_handler = SignalHandlers::alarmHandler;
+        sigemptyset(&action.sa_mask);
+        action.sa_flags = SA_RESTART;
+
+
+        DEBUG_PRINT("this pid is " << getpid() << endl);
+        if (signal(SIGTSTP, SignalHandlers::ctrlZHandler) == SIG_ERR) {
+            perror("smash error: failed to set ctrl-Z handler");
+        }
+        if (signal(SIGINT, SignalHandlers::ctrlCHandler) == SIG_ERR) {
+            perror("smash error: failed to set ctrl-C handler");
+        }
+
+        if(sigaction(SIGALRM , &action, nullptr) == -1) {
+            perror("smash error: failed to set alarm signal handler");
+        }
 
     SmallShell& smash = SmallShell::getInstance();
     shell = &smash;
